Refactor entire components folder
