+++ client/src/components/Home/Center/center.scss
