+++ client/src/components/Home/ExerciseList/ExerciseListItem.scss
