+++ client/src/components/Home/FriendList/FriendListItem.scss
