+++ client/src/components/Home/LeftPanel/leftbar.scss
