+++ client/src/components/Home/RightPanel/rightbar.scss
