+++ client/src/components/Home/index.scss
